+++ src/bindings/python/__init__.py
@@ -2,3 +2,6 @@ try:
     from .python_valhalla import *
 except ModuleNotFoundError:
     from python_valhalla import *
+
+from .actor import Actor
+from .config import get_config