+++ src/thor/bidirectional_astar.cc
@@ -358,8 +358,8 @@ inline bool BidirectionalAStar::ExpandInner(baldr::GraphReader& graphreader,
 
   // setting this edge as reached
   if (expansion_callback_) {
-    expansion_callback_(graphreader, "bidirectional_astar", FORWARD ? meta.edge_id : opp_edge_id, "r",
-                        false);
+    expansion_callback_(graphreader, FORWARD ? meta.edge_id : opp_edge_id, "bidirectional_astar", "r",
+                        pred.cost().secs, pred.path_distance(), pred.cost().cost);
   }
 
   // we've just added this edge to the queue, but we won't expand from it if it's a not-thru edge that
@@ -702,7 +702,8 @@ BidirectionalAStar::GetBestPath(valhalla::Location& origin,
 
       // setting this edge as settled
       if (expansion_callback_) {
-        expansion_callback_(graphreader, "bidirectional_astar", fwd_pred.edgeid(), "s", false);
+        expansion_callback_(graphreader, fwd_pred.edgeid(), "bidirectional_astar", "s",
+                            fwd_pred.cost().secs, fwd_pred.path_distance(), fwd_pred.cost().cost);
       }
 
       // Prune path if predecessor is not a through edge or if the maximum
@@ -745,7 +746,8 @@ BidirectionalAStar::GetBestPath(valhalla::Location& origin,
 
       // setting this edge as settled, sending the opposing because this is the reverse tree
       if (expansion_callback_) {
-        expansion_callback_(graphreader, "bidirectional_astar", rev_pred.opp_edgeid(), "s", false);
+        expansion_callback_(graphreader, rev_pred.opp_edgeid(), "bidirectional_astar", "s",
+                            rev_pred.cost().secs, rev_pred.path_distance(), rev_pred.cost().cost);
       }
 
       // Prune path if predecessor is not a through edge
@@ -856,7 +858,8 @@ bool BidirectionalAStar::SetForwardConnection(GraphReader& graphreader, const BD
 
   // setting this edge as connected
   if (expansion_callback_) {
-    expansion_callback_(graphreader, "bidirectional_astar", pred.edgeid(), "c", false);
+    expansion_callback_(graphreader, pred.edgeid(), "bidirectional_astar", "c", pred.cost().secs,
+                        pred.path_distance(), pred.cost().cost);
   }
 
   return true;
@@ -924,7 +927,8 @@ bool BidirectionalAStar::SetReverseConnection(GraphReader& graphreader, const BD
 
   // setting this edge as connected, sending the opposing because this is the reverse tree
   if (expansion_callback_) {
-    expansion_callback_(graphreader, "bidirectional_astar", fwd_edge_id, "c", false);
+    expansion_callback_(graphreader, fwd_edge_id, "bidirectional_astar", "c", fwd_pred.cost().secs,
+                        fwd_pred.path_distance(), fwd_pred.cost().cost);
   }
 
   return true;
@@ -1004,7 +1008,8 @@ void BidirectionalAStar::SetOrigin(GraphReader& graphreader,
 
     // setting this edge as reached
     if (expansion_callback_) {
-      expansion_callback_(graphreader, "bidirectional_astar", edgeid, "r", false);
+      expansion_callback_(graphreader, edgeid, "bidirectional_astar", "r", cost.secs, edge.distance(),
+                          cost.cost);
     }
 
     // Set the initial not_thru flag to false. There is an issue with not_thru
@@ -1095,9 +1100,10 @@ void BidirectionalAStar::SetDestination(GraphReader& graphreader,
                                      sif::InternalTurn::kNoTurn, -1);
     adjacencylist_reverse_.add(idx);
 
-    // setting this edge as settled, sending the opposing because this is the reverse tree
+    // setting this edge as reached, sending the opposing because this is the reverse tree
     if (expansion_callback_) {
-      expansion_callback_(graphreader, "bidirectional_astar", edgeid, "r", false);
+      expansion_callback_(graphreader, edgeid, "bidirectional_astar", "r", cost.secs, edge.distance(),
+                          cost.cost);
     }
 
     // Set the initial not_thru flag to false. There is an issue with not_thru